+++ src/pygpsclient/signalsview_frame.py
@@ -1,20 +1,23 @@
 """
-levelsview_frame.py
+signalsview_frame.py
 
-Level view frame class for PyGPSClient application.
+Signals view frame class for PyGPSClient application.
 
-This handles a frame containing a graph of current satellite cno levels.
+This handles a frame containing a graph of current signal C/No level,
+correction source and other signal-related flags.
 
-Created on 14 Sep 2020
+Created on 24 Dec 2025
 
 :author: semuadmin (Steve Smith)
 :copyright: 2020 semuadmin
 :license: BSD 3-Clause
 """
 
-# pylint: disable=no-member
+# pylint: disable=no-member, unused-variable, duplicate-code
 
-from tkinter import NE, NSEW, Frame, font
+from tkinter import ALL, NSEW, SE, Frame, N, S, W, font
+
+from pyubx2 import CORRSOURCE, SIGID, UBXMessage
 
 from pygpsclient.canvas_plot import (
     TAG_DATA,
@@ -29,19 +32,40 @@ from pygpsclient.globals import (
     GNSS_LIST,
     GRIDMAJCOL,
     MAX_SNR,
-    WIDGETU2,
+    PNTCOL,
+    SIGNALSVIEW,
+    WIDGETU3,
 )
-from pygpsclient.helpers import col2contrast, fitfont, unused_sats
+from pygpsclient.helpers import col2contrast, fitfont, setubxrate
+from pygpsclient.strings import DLGENABLENAVSIG, DLGNONAVSIG, DLGWAITNAVSIG
 
 OL_WID = 1
 FONTSCALELG = 40
-XLBLANGLE = 35
-XLBLFMT = "000"
+MAXWAIT = 10
+ACTIVE = ""
+XLBLANGLE = 60
+XLBLFMT = "000 WWW_W/W"
+# Correction source legend
+CSLEG = ", ".join(
+    f"{key} {val}" for key, val in CORRSOURCE.items() if key != 0
+).replace(", 7", ",\n7")
 
 
-class LevelsviewFrame(Frame):
+def unused_sigs(data: dict) -> int:
     """
-    Levelsview frame class.
+    Get number of 'unused' sigs in gnss_data.sig_data.
+
+    :param dict data: sig_data
+    :return: number of sigs where cno = 0
+    :rtype: int
+    """
+
+    return sum(1 for (_, _, _, cno, _, _, _, _) in data.values() if cno == 0)
+
+
+class SignalsviewFrame(Frame):
+    """
+    Signalsview frame class.
     """
 
     def __init__(self, app, *args, **kwargs):
@@ -58,10 +82,13 @@ class LevelsviewFrame(Frame):
 
         Frame.__init__(self, self.__master, *args, **kwargs)
 
-        def_w, def_h = WIDGETU2
+        def_w, def_h = WIDGETU3
         self.width = kwargs.get("width", def_w)
         self.height = kwargs.get("height", def_h)
         self._redraw = True
+        self._navsig_status = DLGENABLENAVSIG
+        self._pending_confs = {}
+        self._waits = 0
         self._body()
         self._attach_events()
 
@@ -101,7 +128,7 @@ class LevelsviewFrame(Frame):
 
     def _on_cno0(self, event):  # pylint: disable=unused-argument
         """
-        On double-right-click - include levels where C/No = 0.
+        On double-right-click - include signals where C/No = 0.
 
         :param event: event
         """
@@ -114,6 +141,65 @@ class LevelsviewFrame(Frame):
         )
         self._redraw = True
 
+    def enable_messages(self, status: bool):
+        """
+        Enable/disable UBX NAV-SIG message.
+
+        :param bool status: 0 = off, 1 = on
+        """
+
+        setubxrate(self.__app, "NAV-SIG", status)
+        for msgid in ("ACK-ACK", "ACK-NAK"):
+            self._set_pending(msgid, SIGNALSVIEW)
+        self._navsig_status = DLGWAITNAVSIG
+
+    def _set_pending(self, msgid: int, ubxfrm: int):
+        """
+        Set pending confirmation flag for Signalsview frame to
+        signify that it's waiting for a confirmation message.
+
+        :param int msgid: UBX message identity
+        :param int ubxfrm: integer representing UBX configuration frame
+        """
+
+        self._pending_confs[msgid] = ubxfrm
+
+    def update_pending(self, msg: UBXMessage):
+        """
+        Receives polled confirmation message from the ubx_handler and
+        updates signalsview canvas.
+
+        :param UBXMessage msg: UBX config message
+        """
+
+        pending = self._pending_confs.get(msg.identity, False)
+
+        if pending and msg.identity == "ACK-NAK":
+            self.reset()
+            w, h = self.width, self.height
+            self._canvas.create_text(
+                w / 2,
+                h / 2,
+                text=DLGNONAVSIG,
+                fill=PNTCOL,
+                anchor=S,
+                tags=TAG_DATA,
+            )
+            self._pending_confs.pop("ACK-NAK")
+            self._navsig_status = DLGNONAVSIG
+
+        if self._pending_confs.get("ACK-ACK", False):
+            self._pending_confs.pop("ACK-ACK")
+
+    def reset(self):
+        """
+        Reset spectrumview frame.
+        """
+
+        self.__app.gnss_status.sig_data = []
+        self._canvas.delete(ALL)
+        self.update_frame()
+
     def init_frame(self):
         """
         Initialise graph view
@@ -124,7 +210,7 @@ class LevelsviewFrame(Frame):
         self._canvas.create_graph(
             xdatamax=10,
             ydatamax=(MAX_SNR,),
-            xtickmaj=5,
+            xtickmaj=10,
             ytickmaj=int(MAX_SNR / 10),
             ylegend=("C/No dBHz",),
             ycol=(FGCOL,),
@@ -136,15 +222,25 @@ class LevelsviewFrame(Frame):
         )
         self._redraw = False
 
+        # display 'enable NAV-SIG' warning
+        self._canvas.create_text(
+            self.width / 2,
+            self.height / 2,
+            text=self._navsig_status,
+            fill=PNTCOL,
+            tags=TAG_DATA,
+        )
+
     def _draw_legend(self):
         """
-        Draw GNSS color code legend
+        Draw GNSS color code and correction source legends
         """
 
-        w = self.width / 12
+        w = self.width / 12 / 2
         h = self.height / 18
 
-        lgfont = font.Font(size=int(min(self.width, self.height) / FONTSCALELG))
+        # gnssid color code legend
+        lgfont = font.Font(size=int(min(self.width / 2, self.height) / FONTSCALELG))
         for i, (_, (gnssName, gnssCol)) in enumerate(GNSS_LIST.items()):
             x = (self._canvas.xoffl * 2) + w * i
             self._canvas.create_rectangle(
@@ -166,16 +262,35 @@ class LevelsviewFrame(Frame):
                 tags=TAG_XLABEL,
             )
 
+        # correction source legend
+        self._canvas.create_text(
+            self.width / 2,
+            self._canvas.yofft + h / 2,
+            text=f"Correction Source:\n{CSLEG}",
+            fill=FGCOL,
+            font=lgfont,
+            anchor=W,
+            tags=TAG_DATA,
+        )
+
     def update_frame(self):
         """
-        Plot satellites' signal-to-noise ratio (cno).
+        Plot signal signal-to-noise ratio (C/No).
         Automatically adjust y axis according to number of satellites in view.
         """
 
-        data = self.__app.gnss_status.gsv_data
+        data = self.__app.gnss_status.sig_data
+        if len(data) == 0:
+            if self._waits >= MAXWAIT:
+                self._navsig_status = DLGNONAVSIG
+            else:
+                self._waits += 1
+        else:
+            self._waits = 0
+            self._navsig_status = ACTIVE
         show_unused = self.__app.configuration.get("unusedsat_b")
         siv = len(data)
-        siv = siv if show_unused else siv - unused_sats(data)
+        siv = siv if show_unused else siv - unused_sigs(data)
         if siv <= 0:
             return
 
@@ -184,33 +299,52 @@ class LevelsviewFrame(Frame):
 
         offset = self._canvas.xoffl
         colwidth = (w - self._canvas.xoffl - self._canvas.xoffr + 1) / siv
-        xfnt, _, _ = fitfont(XLBLFMT, colwidth, self._canvas.yoffb, XLBLANGLE)
-        for val in sorted(data.values()):  # sort by ascending gnssid, svid
-            gnssId, prn, _, _, cno, _ = val
+        xfnt, _, _ = fitfont(
+            XLBLFMT,
+            colwidth * 1.66,
+            self._canvas.yoffb,
+            XLBLANGLE,
+        )
+        for val in sorted(data.values()):  # sort by ascending gnssid, svid, sigid
+            gnssId, prn, sigid, cno, corrsource, quality, flags, _ = val
             if cno == 0 and not show_unused:
                 continue
+            sig = SIGID.get((gnssId, sigid), sigid)
             snr_y = int(cno) * (h - self._canvas.yoffb - 1) / MAX_SNR
             (_, ol_col) = GNSS_LIST[gnssId]
+            prn = f"{int(prn):02}"
             self._canvas.create_rectangle(
                 offset,
                 h - self._canvas.yoffb - 1,
                 offset + colwidth - OL_WID,
-                h - self._canvas.yoffb - 1 - snr_y,
+                h - self._canvas.yoffb - snr_y - 1,
                 outline=GRIDMAJCOL,
                 fill=ol_col,
                 width=OL_WID,
                 tags=TAG_DATA,
             )
+            # xlabel prn - sigid
             self._canvas.create_text(
-                offset + colwidth / 2,
-                h - self._canvas.yoffb - 1,
-                text=f"{int(prn):02}",
+                offset + colwidth,
+                h - self._canvas.yoffb + 2,
+                text=f"{prn} {sig}",
                 fill=FGCOL,
                 font=xfnt,
                 angle=XLBLANGLE,
-                anchor=NE,
+                anchor=SE,
                 tags=TAG_DATA,
             )
+            # xcaption corrsource if > 0
+            if corrsource:
+                self._canvas.create_text(
+                    offset + colwidth / 2,
+                    h - self._canvas.yoffb - snr_y + 2,
+                    text=corrsource,
+                    fill=col2contrast(ol_col),
+                    font=xfnt,
+                    anchor=N,
+                    tags=TAG_DATA,
+                )
             offset += colwidth
 
         if self.__app.configuration.get("legend_b"):